Zwei Stellen, an denen sich beides beruehrt:

- Das Raster haelt seit main zwei Byte je Feld (bis 375 Farben). Der
  Vergleich zeichnet aus den vollen Rasterdaten und bekommt denselben
  Typ.
- Der Reiter mit den Staenden heisst jetzt "gemerkt" und liegt in der
  neuen Werkzeugleiste. Der Knopf "Zwei Staende vergleichen" sitzt
  weiterhin ueber der Staendeleiste.

In der README bleiben beide Absaetze: die Fassungsnummer der
Rasterformate von main und die sechs Laeden der Browserdatenbank.

+++ src/components/Vergleich.tsx
@@ -32,7 +32,7 @@ const LUPE_STUFEN = [1, 1.5, 2, 3, 4, 6, 8] as const;
 type Geladen = {
   breite: number;
   hoehe: number;
-  raster: Uint8Array;
+  raster: Uint16Array;
   palette: PalettenEintrag[];
 };
